separate expectations for php 7.4+
